fixed crash due to off-by-one array access in PXPalette_removeColorAtIndex()
After removing the desired color entry (PXPaletteColorPair) from the array by shifting the remaining entries forward, the index used to clear the last entry in the original (pre-shifted) position was off by one (self->colorCount instead of self->colorCount-1).

This cleared the wrong entry (the one immediately after the one that should have been cleared) and left two entries with the same color pointer in the array (even though the second was outside the bounds of self->colorCount). If the PXPalette was later dealloc'd before the second entry was overwritten, the color would be released an extra time and crash (PXPalette_dealloc() uses self->size not self->colorCount as the counter when looping over the array to release colors).

Also, when self->colorCount was equal to self->size, calling PXPalette_remoteColorAtIndex() could corrupt memory, because then self->colors[self->colorCount] was outside the allocated bounds of the array.