PXImage_NSImage(): fixed NSImageCacheException & potential drawing error
Added two error checks:

The first check, before NSImage allocation, returns nil if the image would be zero size (either dimension <= 0). NSImage can successfully init a zero-size instance, but sending a lockFocus message to a zero-size image throws an NSImageCacheException. The exception appears to be caught at the system level, so it won't crash (though a "Can't cache image" message is logged), but it also means any remaining code in the call stack is bypassed, causing other problems. (The canvas becomes frozen & won't allow any more drawing if the arrow keys are used to shift a selection completely off the page - an action that causes a zero-size PXImage.)

The second check returns nil if allocation failed; while unlikely, if this happened, the followup lockFocus message would get sent to nil and have no effect - the focus would stay on the previous view/image - so the call to PXImage_drawRect() would draw in the wrong place.