fixed crash when saving to GIF format
In the PXLayer instance method, adaptToPalette:withTransparency:matteColor:, when the transparency parameter's value is YES (it's currently always called with this value), it's assumed that the bitmapData from the image has 4 channels: In each iteration of the loop over the bitmapData pixels, the current pixel's 4th byte (bitmapData[base + 3]) is accessed.

However, if the bitmapData doesn't have an alpha channel (which happens when all the image's pixels are nontransparent), then each pixel only has 3 bytes. In this case, reading the "4th" byte just gives the first byte of the following pixel. For the last pixel in bitmapData, this is past the end of bitmapData's allocated memory and can cause an EXC_BAD_ACCESS crash.

This patch adds a check for whether bitmapData has an alpha channel - if not, transparency is set to NO, so that the pixels' nonexistent 4th bytes aren't accessed.