Add shift lock while drawing with pencil tool
Mimics the behaviour in Photoshop, with the
exception of where the lock point is set. PS will
set it to the mouse down point, this commit will
set it to the point at which shift is first
pressed. I think this behaviour is more intuitive,
although it breaks with the convention.